+++ Graph/Graph/src/com/gmail/at/ivanehreshi/graph/EdgeTo.java
@@ -2,7 +2,7 @@ package com.gmail.at.ivanehreshi.graph;
 
 public class EdgeTo {
 	public int to;
-	public int w = 0;
+	public double w = 0;
 	
 	public EdgeTo(int to, int w){
 		this.to = to;
